+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.0.2
 
 - Fix potential issue around loading java path on fresh installs
+- Fix legacy resources not installing